+++ tybotRoute/tiledeskChatbotPlugs/directives/Directives.js
@@ -30,6 +30,7 @@ class Directives {
   static GPT_TASK = "gpt_task";
   static FORM = "form";
   static CAPTURE_USER_REPLY = "capture_user_reply";
+  static QAPLA = 'qapla';
 
   // static WHEN_ONLINE_MOVE_TO_AGENT = "whenonlinemovetoagent"; // DEPRECATED?
   // static WHEN_OFFLINE_HOURS = "whenofflinehours"; // DEPRECATED // adds a message on top of the original message when offline hours opts: --replace